add get_qualities xmlrpc function

+++ canape/quality.py
@@ -1,18 +1,18 @@
 #encoding:utf-8
 #       quality.py
-#       
+#
 #       Copyright 2011 nicolas <nicolas@jombi.fr>
-#       
+#
 #       This program is free software; you can redistribute it and/or modify
 #       it under the terms of the GNU General Public License as published by
 #       the Free Software Foundation; either version 2 of the License, or
 #       (at your option) any later version.
-#       
+#
 #       This program is distributed in the hope that it will be useful,
 #       but WITHOUT ANY WARRANTY; without even the implied warranty of
 #       MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
 #       GNU General Public License for more details.
-#       
+#
 #       You should have received a copy of the GNU General Public License
 #       along with this program; if not, write to the Free Software
 #       Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston,
@@ -31,8 +31,8 @@ LOGGER = logging.getLogger(__name__)
 LOCK = Lock()
 
 class Quality(object):
-    
-    def __init__(self, label=None, size=None, extensions=[], 
+
+    def __init__(self, label=None, size=None, extensions=[],
                  keywords=[], extras=[],fromxml=None):
 
         if label == None and fromxml == None:
@@ -47,16 +47,16 @@ class Quality(object):
             self.extensions = extensions
             self.keywords = keywords
             self.extras = extras
-    
+
     def _populate_from_xml(self, fromxml):
         for e in fromxml:
             if e.tag == 'label':
                 setattr(self, 'label', e.text.lower())
             elif e.tag == 'size':
-                setattr(self, 'size', (int(e.attrib['min']), int(e.attrib['max'])))
+                setattr(self, 'size', (float(e.attrib['min']), float(e.attrib['max'])))
             else:
                 setattr(self, e.tag, [child.text.lower() for child in e.iterchildren()])
-    
+
     def _to_xml(self):
         ele_quality = etree.Element("quality")
         etree.SubElement(ele_quality, "label").text = self.label
@@ -65,34 +65,34 @@ class Quality(object):
         for ext in self.extensions:
             etree.SubElement(ele_extensions, "ext").text = ext
         ele_quality.append(ele_extensions)
-        
+
         ele_keywords = etree.Element("keywords")
         for keyword in self.keywords:
             etree.SubElement(ele_keywords, "keyword").text = keyword
         ele_quality.append(ele_keywords)
-        
+
         ele_extras = etree.Element("extras")
         for extra in self.extras:
             etree.SubElement(ele_extras, "extra").text = extra
         ele_quality.append(ele_extras)
         return etree.tostring(ele_quality,pretty_print=True)
-    
+
     def test_quality(self, name, size=None, extension=None):
         scoring = {'label': 50, 'extensions': 30, 'size': 25, 'keywords': 10, 'extras': 5 }
         name = name.lower()
         score = 0
-        
+
         if self.label in name:
             score += scoring['label']
         else:
             score -= scoring['label']
-        
+
         if size is not None:
             if size >= self.size[0] and size <= self.size[1]:
                 score += scoring['size']
             else:
                 score -= scoring['size']
-        
+
         if extension is not None:
             found = False
             for ext in self.extensions:
@@ -101,7 +101,7 @@ class Quality(object):
                     found = True
             if found == False:
                 score -= scoring['extensions']
-            
+
         for keyword in self.keywords:
             if keyword in name:
                 score += scoring['keywords']
@@ -110,10 +110,10 @@ class Quality(object):
             if keyword in name:
                 score += scoring['extras']
         return score
-        
+
     def __repr__(self):
         return self._to_xml()
-    
+
     def __str__(self):
         return self.label
 
@@ -121,7 +121,7 @@ class Qualities(object):
     def __init__(self, qualitiesdb):
         self.db = Qualitiesdb(qualitiesdb)
         self.qualities = list(self.db.iter_qualities())
-    
+
     def compute_scoring(self, name, size=None, extension=None):
         """ Compute qualities score
         return a dict with:
@@ -131,7 +131,7 @@ class Qualities(object):
         for q in self.qualities:
             r[q.label] = q.test_quality(name, size, extension)
         return r
-    
+
     def quality_bet(self, name, size=None, extension=None):
         """ Try to determine the quality
         return a tuple (quality, score)
@@ -142,10 +142,10 @@ class Qualities(object):
         return max(computed.items(), key=lambda q: q[1])
 
 class Qualitiesdb(object):
-    
+
     def __init__(self, xmlfile):
         self.xmlfile = xmlfile
-        
+
     @synchronized(LOCK)
     def add_quality(self, quality):
         tmp_file  = tempfile.NamedTemporaryFile(delete=False)
@@ -160,7 +160,7 @@ class Qualitiesdb(object):
         tmp_file.write('</qualities>')
         tmp_file.close()
         shutil.move(tmp_file.name, self.xmlfile)
-    
+
     @synchronized(LOCK)
     def remove_quality(self, label):
         self.tmpfound=False
@@ -187,7 +187,7 @@ class Qualitiesdb(object):
             LOGGER.error("Can't remove %s because it don't exist in db" % name)
             tmp_file.close()
             os.remove(tmp_file.name)
-        
+
     @synchronized(LOCK)
     def iter_qualities(self):
         context = etree.iterparse(self.xmlfile, events=('end',), tag='quality')
@@ -199,7 +199,7 @@ class Qualitiesdb(object):
             while elem.getprevious() is not None:
                 del elem.getparent()[0]
         del context
-        
+
     def _fast_iter(self, function):
         context = etree.iterparse(self.xmlfile, events=('end',), tag='quality')
         for event, elem in context:
@@ -221,4 +221,4 @@ if __name__ == '__main__':
     for r in results:
         print "Name: %s sourcescore: %s  quality score: %s" % (r.name, r.sourcescore,
             qualities.quality_bet(r.name,size=r.size))
-        
+
